requirements.txt
flask
gunicorn
mysql-connector-python
python-dotenv
twilio
openai